+++ src/main/java/com/foolsix/fancyenchantments/events/CapabilityEvents.java
@@ -64,7 +64,12 @@ public class CapabilityEvents {
                 if (elementStats.getPoint(AER) >= CONFIG.aerLevelToGetSpeed) {
                     e.player.addEffect(new MobEffectInstance(MobEffects.MOVEMENT_SPEED,
                             20,
-                            elementStats.getPoint(AER) / CONFIG.aerLevelToGetSpeed - 1));
+                            Math.min(CONFIG.maxAerEffectLevel, elementStats.getPoint(AER) / CONFIG.aerLevelToGetSpeed - 1)));
+                }
+                if (elementStats.getPoint(IGNIS) >= CONFIG.ignisLevelToGetStrength) {
+                    e.player.addEffect(new MobEffectInstance(MobEffects.DAMAGE_BOOST,
+                            20,
+                            Math.min(CONFIG.maxIgnisEffectLevel, elementStats.getPoint(IGNIS) / CONFIG.ignisLevelToGetStrength - 1)));
                 }
                 if (elementStats.getPoint(IGNIS) >= CONFIG.ignisLevelToGetFireResistance) {
                     e.player.addEffect(new MobEffectInstance(MobEffects.FIRE_RESISTANCE, 20));
@@ -72,12 +77,12 @@ public class CapabilityEvents {
                 if (elementStats.getPoint(TERRA) >= CONFIG.terraLevelToGetResistance) {
                     e.player.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE,
                             20,
-                            elementStats.getPoint(TERRA) / CONFIG.terraLevelToGetResistance - 1));
+                            Math.min(CONFIG.maxTerraEffectLevel, elementStats.getPoint(TERRA) / CONFIG.terraLevelToGetResistance - 1)));
                 }
                 if (elementStats.getPoint(AQUA) >= CONFIG.aquaLevelToGetRegeneration) {
                     e.player.addEffect(new MobEffectInstance(MobEffects.REGENERATION,
                             20,
-                            elementStats.getPoint(AQUA) / CONFIG.aquaLevelToGetRegeneration - 1));
+                            Math.min(CONFIG.maxAquaEffectLevel, elementStats.getPoint(AQUA) / CONFIG.aquaLevelToGetRegeneration - 1)));
                 }
                 if (elementStats.getPoint(TWISTED) - elementStats.getPoint(HOLY) >= CONFIG.twistedLevelToGetDebuff) {
                     if (Math.random() < CONFIG.probabilityToGetDebuff) {